Create Break and continue
The program is just a simple example to understand how to break and continue keywords work. Here break will cause to go out from the for loop once the value of i reaches 3, on the other hand, the continuing example will just skip the value of 3 when i reaches 3 and will continue printing the number staying inside the loop until for reaches the threshold. 
( the example is based on the cisco python essentials program)
=> The next example is all about break and continue using while true.